+++ services/src/main/java/org/keycloak/connections/httpclient/DefaultHttpClientFactory.java
@@ -134,8 +134,8 @@ public class DefaultHttpClientFactory implements HttpClientFactory {
         if (httpClient == null) {
             synchronized(this) {
                 if (httpClient == null) {
-                    int socketTimeout = config.getInt("socket-timeout-millis", 5000);
-                    int establishConnectionTimeout = config.getInt("establish-connection-timeout-millis", -1);
+                    long socketTimeout = config.getLong("socket-timeout-millis", 5000L);
+                    long establishConnectionTimeout = config.getLong("establish-connection-timeout-millis", -1L);
                     int maxPooledPerRoute = config.getInt("max-pooled-per-route", 64);
                     int connectionPoolSize = config.getInt("connection-pool-size", 128);
                     long connectionTTL = config.getLong("connection-ttl-millis", -1L);
